Start of real loadout sharing
